Added watermark to all figures and USYD logo to about page (#69)
* Added watermark to all figures and USYD logo to about page

The watermark font size is scaled based upon the number of figures in each row of the website. This is based upon the lg argument of the dcc.Col. For now the watermark uses a static web address.

The USYD logo on the about page links to the USYD business school website.

* Reformatted util.py and pages.py