Watermark and usyd logo (#70)
* Added watermark to all figures and USYD logo to about page

The watermark font size is scaled based upon the number of figures in each row of the website. This is based upon the lg argument of the dcc.Col. For now the watermark uses a static web address.

The USYD logo on the about page links to the USYD business school website.

* Reformatted util.py and pages.py

* Fixed the missed USYD logo on the about from last commit

* Added USYD logo .png file to the assets folder