allow users to download data as .xlsx file with multiple sheets
Added functionality to download the forecast data, cross validation scores table and the raw series data as a .xlsx file. Each aformentioned dataset is a given a seperate sheet in the excel file.

The current implementation of this functionality uses the io and base64 packages (see https://community.plotly.com/t/allow-users-to-dowload-an-excel-in-a-click/9410/15).