Scoring function fix (#92)
* Add MASE and Winkler scoring functions and associated functionality.

Add MASE and Winkler scores (for confidence intervals) to a cv_scores dictionary. Plot download button and forecast data download link on series pages. Replace forecast table with CV scores table on series pages. Only the 95% CI Winkler score is reported on the series page CV scores table. The "best model" functionality currently only uses MSE.

* re-format using pre_push.sh

* Remove comments from other forecast models and multiprocessing

* update the CV scores table to a sortable dash DataTable, download plot as svg not png

* re-format using pre_push.sh

* allow users to download data as .xlsx file with multiple sheets

Added functionality to download the forecast data, cross validation scores table and the raw series data as a .xlsx file. Each aformentioned dataset is a given a seperate sheet in the excel file.

The current implementation of this functionality uses the io and base64 packages (see https://community.plotly.com/t/allow-users-to-dowload-an-excel-in-a-click/9410/15).

* Allow users  to select the scoring function for leaderboard

The leaderboard page now includes a dropdown menu for users to select a scoring function to rank forecast methods over all series.

The default scoring function is MSE.

* Restore all data sources and models

* Fixed formatting

* Fixed bootstrap compatibility issue

* Fix CV table bug and remove print statements

Co-authored-by: Robert James <rjames6023@gmail.com>