Fix sqrt issue (#91) and add tooltips (#90)
The seasonality test sqrt issue was caused by not squaring the ACF function. In the python version of the M4 benchmark code. The R version of the M4 benchmark code does square the ACF function and this is mentioned in the M4 python code benchmark file.

Tooltips have been added to describe the scoring functions.