Improve find series filters (#158)
* Add mailto for email

* Add function to create metadata #110

This is to address #110.

* add model versioning #131

Keep a version of the downloaded file. If no new downloads, then model does not need updating.

* run black

* yearly forecasts

Forecasting for yearly data is 4 years ahead at the moment.

* add version and run black

* Update pages.py

* Fix the link for #148

The navbar is now collapsible on smaller screens.

* Fix CV table Sorting #134

Buttons also added to switch between relative scores and raw scores for the CV table

* Add Downloader For ABS Data #140

* initial upload for #139

Generate thumbnails separately and rend as cards.

* Update pages.py

* Create .gitignore

* add sorting series

* clean comments

* run black

* rebase and clean

issue #139

* run black

* responsive layout

Change layout sizes for different screen sizes.  helps with issue #57

* Amend to support multiple pages

* update for new dash layout #158

* update model versioning

* run black

* single search

* fix brackets

* run black

* Update search.py

* clean file

Co-authored-by: Ardalan Mirzaei <ardalan.mirzaei@sydney.edu.au>
Co-authored-by: Stephen Tierney <sjtrny@gmail.com>